account-management-api/src/test/java/uk/gov/di/accountmanagement/lambda/RetrieveMfaMethodsHandlerTest.java
package uk.gov.di.accountmanagement.lambda;

import com.amazonaws.services.lambda.runtime.Context;
import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.mock;

public class RetrieveMfaMethodsHandlerTest {
    @Test
    void shouldReturn200ForSuccessfulRequest() {
        var handler = new RetrieveMfaMethodsHandler();
        var context = mock(Context.class);
        var event = new APIGatewayProxyRequestEvent();

        var result = handler.handleRequest(event, context);
        assertEquals(200, result.getStatusCode());
    }
}
